Cleaned and split routing
